feat: handle device event in service handler
Add device event in service control to allow the user to receive and
handle device events.
The data of those events is for the moment not parsed. Converting those
events to Rust is not trivial, and we can let the user do it rather than
doing it in his place, when he does not necessarily care for all the
different type of device events.